+++ Phoenix/NSScreen+PHExtension.m
@@ -88,7 +88,7 @@ static NSString * const NSScreenNumberKey = @"NSScreenNumber";
     NSPredicate *spaceIsOnThisScreen = [NSPredicate predicateWithBlock:
                                         ^BOOL (PHSpace *space, __unused NSDictionary<NSString *, id> *bindings) {
 
-        return [[space screen] isEqualTo:self];
+        return [[space screens] containsObject:self];
     }];
 
     return [[PHSpace all] filteredArrayUsingPredicate:spaceIsOnThisScreen];